update bspwmrc
Move desktop creation down, because there was an issue with xfce4
panel's workspace displayer. It is initialized in the host specific
files, but it does not matter if it is done later.

Turn of monitor managaging by bspwm for now because of bugs in scripts.

Change gap and remove some hidpi bs.